+++ front/src/components/Header.js
@@ -1,27 +1,40 @@
 import React from 'react';
+import { useState, useCallback } from 'react';
 import Img from '../assets/codong.png';
 
+
 function Header() {
+  const [isHovering, setIsHovering] = useState("");
+
+  const handleMouseHover = useCallback(() => {
+    setIsHovering((isHovering) => !isHovering);
+  }, [])
+
   return (
-    <div style={HeaderStyle}>
-      <div style={TitleStyle}>
-        <b>숭컴타</b>
+    <div>
+      <div style={HeaderStyle}>
+        <div style={TitleStyle}>
+          <b>숭컴타</b>
+        </div>
+        <div style={MenuStyle}
+          onMouseEnter={handleMouseHover}
+          onMouseLeave={handleMouseHover} >
+          <div style={MenuItemStyle}>대결하기</div>
+          <div style={MenuItemStyle}>연습하기</div>
+          <div style={MenuItemStyle}>스크립트</div>
+          <div style={MenuItemStyle}>내 정보</div>
+        </div>
+        <div>
+          <img width='40px' height='40px' src={Img} />
+        </div>
       </div>
-      <div style={MenuStyle}>
-        <div style={MenuItemStyle}>대결하기</div>
-        <div style={MenuItemStyle}>연습하기</div>
-        <div style={MenuItemStyle}>스크립트</div>
-        <div style={MenuItemStyle}>내 정보</div>
-      </div>
-      <div>
-        <img width='40px' height='40px' src={Img} />
-      </div>
+      {isHovering && <div id="detail" style={DetailMenuStyle}>
+      </div> }
     </div>
   )
 }
 
 const HeaderStyle = {
-  position: 'absolute',
   width: '100%',
   overflow: 'hidden',
   background: '#7BC5C5',
@@ -31,19 +44,19 @@ const HeaderStyle = {
   fontFamily: 'Noto Serif KR',
   textAlign: 'left',
   alignItems: 'center',
-  display: 'flex', 
+  display: 'flex',
 }
 
 const TitleStyle = {
   marginLeft: '5%',
-  marginRight : '20%',
-  color : 'white',
+  marginRight: '20%',
+  color: 'white',
   fontSize: '40px'
 }
 
 const MenuStyle = {
   width: '40%',
-  display: 'flex', 
+  display: 'flex',
   alignItems: 'center',
 }
 
@@ -54,4 +67,13 @@ const MenuItemStyle = {
   fontSize: '20px',
 }
 
+const DetailMenuStyle = {
+  width: '100%',
+  background: '#FFFFFF',
+  height: '175px',
+  fontSize: '25px',
+  fontWeight: '400',
+  fontFamily: 'Noto Serif KR',
+  textAlign: 'left',
+}
 export default Header;
